Aplicar arquitetura MSC para o Mongoose
