+++ src/users/usersModel.js
@@ -1,27 +1,43 @@
-const User = require('./schemas');
+const mongoose = require('mongoose');
+const cpfValidate = require('../utils/cpfValidate');
 
-const createUser = async (data) => {
-  const user = new User(data);
-  return user.save();
-}
+const Schema = mongoose.Schema;
 
-const updateUserByPhone = async (data) => {
-  const { phone } = data;
-  const updateThat = Object.entries(data);
-  try {
-    updateThat.forEach(([key, value]) => {
-      User.updateOne({ phone }, { $set: { [key]: value } });
-    });
-    return { modified: 1, items: updateThat.length }
-  } catch (err) {
-    console.error(err);
+const userSchema = new Schema({
+  name: {
+    type: String,
+    required: true,
+  },
+  cpf: {
+    type: String,
+    minlength: 11,
+    maxlength: 14,
+    validate: [cpfValidate, 'Invalid CPF'],
+    unique: true,
+  },
+  phone: {
+    type: String,
+    required: true,
+    unique: true,
+  },
+  address: {
+    cep: {
+      type: String,
+      required: true
+    },
+    number: {
+      type: String,
+      required: true
+    },
+    complement: String,
+  },
+  purchaseHistory: {
+    type: Array,
+  },
+  desatived: {
+    type: Boolean,
+    default: false
   }
+});
 
-}
-
-const desativeUserByPhone = async(phone) => {
-  const user = User.updateOne({ phone }, { $set: { desatived: true } });
-  return user.exec();
-}
-
-module.exports = { createUser, updateUserByPhone, desativeUserByPhone };
+module.exports = mongoose.model("Users", userSchema);
